entreprise

+++ src/main/java/com/example/appwebbellac/model/Entreprise.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Entreprise {
 
-    private long IdEnt;
+    private Integer IdEnt;
 
 
     private String nomEnt;
